requirements.txt
requests
purepng
